requirements.txt
streamlit==1.36.0
psycopg[binary]==3.2.9
pandas==2.2.2
python-dotenv==1.0.1
plotly==5.22.0
